+++ Makefile
@@ -16,7 +16,9 @@ list:
 
 readme:
 	pip install md-toc
-	md_toc -p README.md github --header-levels 4
+	pip install referencer
+	referencer $(PACKAGE) README.md --in-place
+	md_toc -p README.md github --header-levels 3
 	sed -i '/(#$(PACKAGE)-py)/,+2d' README.md
 
 release:
