+++ goodtables/__main__.py
@@ -1,0 +1,7 @@
+from .cli import cli
+
+
+# Module API
+
+if __name__ == "__main__":
+    cli()
